add route Product for asdmin and user

+++ src/services/productService.js
@@ -1,4 +1,5 @@
 const Product = require('../models/productModel')
+const OrderItem = require('../models/orderItemModel')
 
 const getAllProductService = async () => {
     try {
@@ -17,7 +18,7 @@ const getAllProductService = async () => {
     }
 }
 
-const createProductService = async (name, price, imgUrls, description, category, quantity, discount) => {
+const createProductService = async (name, price, imgUrls, description, idCategory, quantity, discount) => {
     try {
         const product = await Product.findOne({ name });
         if (product){
@@ -32,7 +33,7 @@ const createProductService = async (name, price, imgUrls, description, category,
             price: price,
             imgUrls: imgUrls,
             description: description,
-            category: category,
+            idCategory: idCategory,
             quantity: quantity,
             discount: discount,
         })
@@ -83,6 +84,13 @@ const updateProductService = async (_id, name, price, imgUrls, description, cate
 
 const deleteProductService = async (_id) => {
     try {
+        const orderItems = await OrderItem.find({idProduct: _id});
+        if (orderItems.length > 0) {
+           return {
+               EC: 1,
+               EM: "Không thể xóa sản phẩm khi vẫn còn đơn hàng",
+           };
+        }
         const result = await Product.deleteOne({_id: _id})
         if (result.deletedCount === 0) {
             return {
@@ -104,6 +112,32 @@ const deleteProductService = async (_id) => {
     }
 }
 
+const updateAvailableProductsService = async (_id, isActive) => {
+    try{
+        const result = await Product.findOneAndUpdate(
+            {_id: _id},
+            {
+                $set: {
+                    isActive: isActive
+                }
+            },
+            {new: true}
+        )
+        return {
+            EC: 0,
+            EM: "Cập nhật trạng thái bán thành công",
+            data: result,
+        };
+    }catch (error) {
+        return {
+            EC: 1,
+            EM: "Không thể cập nhật trạng thái bán",
+            data: [],
+        };
+    }
+}
+
 module.exports = {
-    getAllProductService, createProductService, updateProductService, deleteProductService
+    getAllProductService, createProductService, updateProductService, deleteProductService,
+    updateAvailableProductsService
 }
